extract num-bigint and bigdecimal interop code into separate modules
